split peer tracker
