+++ client/src/main.rs
@@ -1,109 +1,13 @@
 use common::*;
-use std::collections::HashMap;
-use std::net::{SocketAddrV4, TcpStream};
+use std::net::SocketAddrV4;
 use std::path::PathBuf;
 use std::sync::Arc;
 
 mod file_server;
-use file_server::FileServer;
+use file_server::{FSRequest, FileServer};
 
-use self::file_server::{FileSystem, SimpleFileSystem, FSRequest};
-
-#[derive(Debug, Clone)]
-struct Peer {
-    sock: SocketAddrV4,
-    files: Vec<File>,
-}
-
-#[derive(Default)]
-struct Peers {
-    full: HashMap<SocketAddrV4, Peer>,
-}
-
-impl Peers {
-    fn new() -> Self {
-        Self::default()
-    }
-    fn add_peer(&mut self, peer: Peer) -> Option<Peer> {
-        self.full.insert(peer.sock.clone(), peer)
-    }
-    fn update_peer(&mut self, new_peer: Peer) {
-        self.full.insert(new_peer.sock.clone(), new_peer);
-    }
-    fn remove_peer(&mut self, sock: SocketAddrV4) -> Option<Peer> {
-        self.full.remove(&sock)
-    }
-    fn get_peer(&mut self, sock: SocketAddrV4) -> Option<&Peer> {
-        self.full.get(&sock)
-    }
-}
-
-impl From<server::RegisterPeer> for Peer {
-    fn from(server::RegisterPeer { sock, file_list }: server::RegisterPeer) -> Self {
-        Peer {
-            sock,
-            files: file_list,
-        }
-    }
-}
-
-impl From<server::UpdatePeer> for Peer {
-    fn from(server::UpdatePeer { sock, file_list }: server::UpdatePeer) -> Self {
-        Peer {
-            sock,
-            files: file_list,
-        }
-    }
-}
-
-struct TrackerServerContext<FS: FileSystem> {
-    peers: Peers,
-    server: TcpStream,
-    file_server: Arc<FileServer<FS>>,
-}
-
-impl<FS: FileSystem> TrackerServerContext<FS> {
-    fn handle_message(&mut self, msg: AnyMessage) {
-        match msg {
-            AnyMessage::Server(server::Message::RegisterPeer(p)) => {
-                self.peers.add_peer(p.into());
-            }
-            AnyMessage::Server(server::Message::UpdatePeer(p)) => {
-                self.peers.update_peer(p.into());
-            }
-            AnyMessage::Server(server::Message::UnregisterPeer(p)) => {
-                self.peers.remove_peer(p.sock).unwrap();
-            }
-            AnyMessage::Client(client::Message::RequestFile(f)) => {
-                todo!("Serve file to {f:?}")
-            }
-            m => panic!("can't handle msg {m:?}"),
-        }
-    }
-
-    fn new(srv: TcpStream, fsrv: &Arc<FileServer<FS>>) -> Result<Self, std::io::Error> {
-        let file_server = Arc::clone(&fsrv);
-        srv.set_nonblocking(true)?;
-        let mut slf = Self {
-            peers: Peers::new(),
-            server: srv,
-            file_server,
-        };
-        let connect_msg = client::Message::Connect(client::Connect {
-            serve_port: fsrv.server.local_addr()?.port(),
-            file_list: slf.file_server.file_system.list_files(),
-        });
-        write_msg(&mut slf.server, connect_msg).unwrap();
-        Ok(slf)
-    }
-
-    fn check_server_messages(&mut self) -> Result<(), std::io::Error> {
-        if let Some(m) = read_msg_nb(&mut self.server)? {
-            self.handle_message(m);
-        }
-        Ok(())
-    }
-}
+mod tracker;
+use tracker::TrackerServerContext;
 
 fn main() -> Result<(), std::io::Error> {
     if std::env::args().nth(1) == Some("get".to_string()) {
@@ -115,7 +19,9 @@ fn main() -> Result<(), std::io::Error> {
 
 fn get_file_main() -> Result<(), std::io::Error> {
     use std::io::Read;
-    let mut s = std::net::TcpStream::connect("127.0.0.1:42065").unwrap();
+    let file_server_addr: SocketAddrV4 = "127.0.0.1:6969".parse().unwrap();
+
+    let mut s = std::net::TcpStream::connect(file_server_addr).unwrap();
     let req_msg = client::Message::RequestFile(client::RequestFile {
         file: PathBuf::from("uwu.txt"),
     });
@@ -127,9 +33,13 @@ fn get_file_main() -> Result<(), std::io::Error> {
 }
 
 fn serve_file_main() -> Result<(), std::io::Error> {
-    let track_server = std::net::TcpStream::connect("127.0.0.1:6969").unwrap();
-    let file_ctx = Arc::new(FileServer::<SimpleFileSystem>::new("127.0.0.1:4545".parse().unwrap())?);
-    let mut track_ctx = TrackerServerContext::new(track_server, &file_ctx)?;
+    let tracker_addr = "127.0.0.1:42069".parse().unwrap();
+    let file_server_addr = "127.0.0.1:6969".parse().unwrap();
+
+    let file_ctx = Arc::new(FileServer::<file_server::SimpleFileSystem>::new(
+        file_server_addr,
+    )?);
+    let mut track_ctx = TrackerServerContext::new(tracker_addr, &file_ctx)?;
 
     std::thread::scope(|s| {
         loop {
